feat: mod variable
